Moved onboarding_utils
